Corrected bug in clearBit
The default implementation of clearBit also clears
the bits to the left of the one passed as
parameter. The solution is to give an explicit
implementation.

Also, a quickCheck for clearBit has been added.

Review by Iago Abal <mail@iagoabal.eu>:
---------------------------------------
I personally think that the problem here is that certain operations like
.&. should sign-extend their operands rather than zero-extend. However,
that would be backwards incompatible, and I think this patch is a good
temporary fix. Long term, we should fix the real problem even if that
means changing the philosophy of the library a little bit.
---------------------------------------